requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
bs4
pypdf
arxiv
PyMuPDF
faiss-cpu
langchain_huggingface
scikit-learn
langchain-google-genai
langchain-pinecone
rank_bm25
wikipedia
youtube_search
langgraph
chromadb
duckduckgo-search
llama-index-llms-huggingface-api 
llama-index-embeddings-huggingface

